Fixup for Python 3.10
This commit contains some fixups for Python 3.10.  Apparently some of
the PyQt overloads disappeared that support int <-> float
interchangeably as args to C++ functions, which caused the app to not
work at all on Python 3.10.

Note: there are still extant bugs with using Python 3.10, but this
commit at least allows the app to (mostly) run ok.

+++ electroncash_gui/qt/paytoedit.py
@@ -68,7 +68,7 @@ class PayToEdit(PrintError, ScanQRTextEdit):
         documentMargin = document.documentMargin()
         self.verticalMargins = margins.top() + margins.bottom()
         self.verticalMargins += self.frameWidth() * 2
-        self.verticalMargins += documentMargin * 2
+        self.verticalMargins += int(documentMargin * 2)
 
         self.heightMin = self.fontSpacing + self.verticalMargins
         self.heightMax = (self.fontSpacing * 10) + self.verticalMargins
