+++ electroncash_gui/qt/qrreader/video_overlay.py
@@ -66,7 +66,7 @@ class QrReaderVideoOverlay(QWidget):
         self.bg_rect_pen = QPen()
         self.bg_rect_pen.setColor(Qt.black)
         self.bg_rect_pen.setStyle(Qt.DotLine)
-        self.bg_rect_fill = QColor(255, 255, 255, 255 * self.BG_RECT_OPACITY)
+        self.bg_rect_fill = QColor(255, 255, 255, int(255 * self.BG_RECT_OPACITY))
 
         self.qr_finder = QSvgRenderer(":icons/qr_finder.svg")
